Added signal handlers
Make sure that the child and interrupt signals are handled. The
child signal is recieved when a child process dies. We need to get
the return value of that process so that it does not become a zombie.
Also implement a INTSIG handler so that we can clean up after the
the process is about to terminate.