Added first version of saving the stats
This is the first version of the stat saving. We need to implement
the mutex so that two processes don't write to the database at the
same time. We also need to create an identify command so that the
client can tell the server its name and id. Not sure how the protocol
is going to work just yet. Need to talk to Hardik and Zack. They will
have some good insight on what is the best way to implement this.