Added START command
The START command is sent by the server to the clients to let them
know that the game has started and that the player1 should be able
to move now. Once a client recieves this command it now means that
they are talking to the gameserver and that the gameserver protocol
is not in effect.