Added START command to recieve protocol
Added the START command tot he recieve protocol so that we are able
to actually play the game (if a command is not reconized the recieve
interface throws an exception which caused the program to crash).
Maybe we will add parameters to the start command indicating whos
turn it is.