Added comments and SEM_UNDO
SEM_UNDO tell the kernel that if the process were to die without
having the chance to increment the semaphore (recieves SIGKILL)
then the kernel will undo anything that the process has done to
the semaphore. Just being safe. Also I added comments to make the
code a little bit more readable.