Fixed bugs in semaphore and made it more robust
Fixed a bug where we are init the semaphore and not telling semop
what number semaphore we wanted. There was also a bug where no
matter what the semaphore_init() function would fail. Just make sure
that we return 0 when we know that we have successfully create a
semaphore. I also made the semaphore_wait() and semaphore_signal()
operations return an integer to tell the calling program whether
or not they suceeded. This makes the program a little bit more
robust.