Added semaphores to the driver
Added the wait() and signal() operations to the driver so that we
are sure that when two processes try to access the database at the
same time only one will be allowed in while the other process(es)
wait. Need to test the race condition to ensure that is works.