Added the ID command
The id command is used to idenify the player to the server. Right
now it accepts the player id, first name and last name. These values
are used to store the stats. There are a couple problems that I have
with this implementation. First off the gameserver has to know about
a player record which seems bad. The gameserver should have no knowledge
about statics (unless we want to integrate them). Second the player has
to give the server the player id. This in reality makes no sense since
the client will not have any knowledge about what ids are taken and
which ids are free. And finally it doesn't carry over to the chat
server. Maybe I should have a different authorization and auth
protocol but this will have todo for now.